requirements_desktop.txt
pywebview>=4.0
pyinstaller>=5.0
uvicorn>=0.15.0
fastapi>=0.68.0
sqlalchemy>=1.4.23
pydantic[email]>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
bcrypt>=3.2.0
